Electronic_medical_record_detection_3.1/requirements.txt
Flask==2.0.1
pandas>=1.0.0
openpyxl==3.0.9
python-dotenv==0.19.1
WTForms==3.0.0
Flask-WTF==1.0.0
mysql-connector-python>=8.0.0
transformers>=4.30.0
torch>=2.0.0
spacy>=3.5.0 